Make sure that refreshToken() for the implicit grant flow is an actual noop function
By exporting a noop function in the `auth` module we can make sure that the
refreshToken implementation for the implicit grant flow isn't doing anything.

The current test was simply assuming that a returned value of false was
the signature of a noop function which isn't quite right as we could implement
a refreshToken method that can both do stuff and return false.